feat: automatic support to use client auth sock on any git ops
This piggybacks on Erik's work to isolate sessions safely.
This introduces automatic support of client auth sock to any git ops.

We rely on the SSH_AUTH_SOCK env variable on the host to locate the
socket. This env variable is automatically set when in use of a socket
or with forwarded ssh connection

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>