chore: extend engine.Client implementation to allow scoped sshAuthSockPath
Follow up from previous commit, whose changes are better kept seperated as it
modifies our engine.Client implementation to allow a new way to pass the
sshAuthSockPath, without relying on the host env var.

The extended way to pass the socketPath is to rely on the current engine.ContextWithClientMetatada which enables us to wrap an existing context containing a clientMetadata, which the client initialization grabs and adds as the sshAuthSockPath.

Reasons:
We want to be able to have a fine grained control of the success or
failure of refs to private repos. Relying on host env variables seem to
be dangerous, and a declarative model seem to be more scalable

Also, most of the tests are parallelized and not having to deal with
state concistency across parralel tests sets a better foundation

Now, testOnMultipleVCS modifies the context on-the-fly for tests. The
helper is both expressive and easy to use.

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>