Extend test suite to better isolate tests around SSH_AUTH_SOCK support
Reasons:
We want to be able to have a fine grained control of the success or
failure of refs to private repos. Relying on host env variables seem to
be dangerous, and a declarative model seem to be more scalable

Also, most of the tests are parallelized and not having to deal with
state concistency across parralel tests sets a better foundation

Now, testOnMultipleVCS modifies the context on-the-fly for tests. The
helper is both expressive and easy to use.

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>