chore: update mountedPrivateRepoSocket
mountedPrivateRepoSocket is the helper function to generate an ssh
client per test. As per my tests, github runners have a socket limit of 100k+ units

Due to some race conditions, I removed the t.cleanup towards a defer
cleanup with an initialization in the most nested test, to ensure a
short lifetime.

t.tempdir() is not used in setupPrivateRepoSSHAgent as we were hitting
the limit of path length > 108 bytes on GitHub runners.

Also, one test ("unpinned gets pinned") was updated because the logic did not take into account
the fact that SSH like ref styles could have several "@", as a user
could be introduced

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>