chore: expose GitModuleSource.HtmlRepoURL
GitModuleSource has the htmlURL field, which is the base URL to access the module on a Git provider. In order to do so, it constructs opinionated URLs (with src, tree..). However, this is always an URL linking to a specific module, not the root URL for that repo.

This commit introduces htmlRepoURL: it is the HtmlURL to the root of a Git repo, in which a module is stored

Reason:
Daggerverse currently relies on the user ref, when publishing a module, to transform all relative paths inside all Markdown (which also includes our functions, object documentations / Readme [in the past]). For that, it needs another format of URL, to access the raw content for each provider: https://raw.githubusercontent.com/dagger/dagger/6249cc93f9467434215d155e6e768431e507feb7/core/integration/git_test.go, which can be accessed originally as https://github.com/dagger/dagger/raw/6249cc93f9467434215d155e6e768431e507feb7/core/integration/git_test.go

With the new ref scheme, it is easier to just expose this field

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>