feat: add a "debug" boolean option inside the dagger.json's SDK field
In order to improve and formalize jumping into the context of the
runtime, this commit adds an option to the dagger.json (easiest option)
to jump into the container runtime.

It's a good DX loop to work on the auth:

1. Just add a `debug: true` field -->

```json
{
  "name": "dagger-dev",
  "engineVersion": "v0.19.4",
  "sdk": {
    "source": "go",
    "debug": true
  }
}
```

2. Then, call any dagger function (any return types will work). You will
   jump into the runtime container of that function (Go, Ts and Python
SDKs), prior the call

Important note: As we now lazy load the runtime, this leads to annoying cache
hits (not being able to jump into the container). In order to avoid
that, when this option is set, we burst the cache every time

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>